requirements.txt
fastapi
uvicorn[standard]
pylint
boto3
invoke
annoy
python-multipart
httpx
Pillow==8.1.2
